+++ sdk/eventgrid/azure-messaging-eventgrid/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.19.0-beta.1 (Unreleased)
 
 ### Features Added
+- new ACS Router systemevents.
 
 ### Breaking Changes